changed actions tensor type and added new files to fix a3c train code

+++ _2023_04/_05_A3C/c_actor_and_critic.py
@@ -117,7 +117,7 @@ class Buffer:
 
         # Convert to tensor
         observations = torch.tensor(observations, dtype=torch.float32, device=DEVICE)
-        actions = torch.tensor(actions, dtype=torch.int64, device=DEVICE)
+        actions = torch.tensor(actions, dtype=torch.float32, device=DEVICE)
         next_observations = torch.tensor(next_observations, dtype=torch.float32, device=DEVICE)
         rewards = torch.tensor(rewards, dtype=torch.float32, device=DEVICE)
         dones = torch.tensor(dones, dtype=torch.bool, device=DEVICE)
